requirements.txt
opensearch-py==2.7.1
boto3==1.37.1
requests-aws4auth==1.3
PyYAML==6.0.2
requests==2.32.3
streamlit==1.41.1
aioboto3==14.1.0
aws-cdk-lib==2.186.0
constructs>=10.0.0,<11.0.0